fix(insurance): declare a Number attribute decimal, as PcmGenerator already does
The two maps that declare attributes into the same model disagreed about
Number: the insurance map said int, PcmGenerator's dataTypeToCmlType said
decimal. A Salesforce Number attribute can carry decimal places and the
safe-literal guard already admits 500.5 for one, so decimal is the side to
keep; the doc comment claiming PcmGenerator "normalizes the same way" was
simply wrong and is corrected to say what actually has to hold.

No emitted CML changes in either live path. Value emission is untouched -
isSafeUnquotedLiteral treats int and decimal identically and the shared
emitter quotes only for `string` - and neither merge path declares
attributes. What changes is the attribute declaration build mode would emit
(no command reaches it; both convert subclasses force merge), and the
underwriting merge's attribute-presence probe, which greps for
`<cmlType> <name>` and so was reporting every curated `decimal` Number
attribute as absent.

+++ src/shared/insurance/insurance-rule-generator.ts
@@ -27,11 +27,19 @@ import {
 
 /**
  * Source data type -> CML data type. Keyed uppercase because payloads are inconsistent about case
- * ('Datetime' and 'DateTime' both occur in real RuleDefinitions), and because PcmGenerator — which
- * declares these same attributes in the same model — normalizes the same way.
+ * ('Datetime' and 'DateTime' both occur in real RuleDefinitions) — the same normalization
+ * PcmGenerator's `dataTypeToCmlType` applies to its own keys.
+ *
+ * PcmGenerator declares these same attributes into the same model, so wherever both maps hold a
+ * key they must resolve it identically or one attribute ends up with two contradictory
+ * declarations. They do not otherwise mirror each other: this map additionally carries the
+ * spellings only a condition payload or an AttributePicklist produces (INTEGER, DECIMAL, DOUBLE,
+ * BOOLEAN, DATETIME), for which PcmGenerator has no entry.
  */
 const SOURCE_DATA_TYPE_TO_CML: Record<string, string> = {
-  NUMBER: CML_DATA_TYPES.INTEGER,
+  // decimal, not int: a Salesforce Number attribute may carry decimal places, and PcmGenerator
+  // declares one `decimal` in this same model.
+  NUMBER: CML_DATA_TYPES.DECIMAL,
   INTEGER: CML_DATA_TYPES.INTEGER,
   PERCENT: CML_DATA_TYPES.DECIMAL,
   CURRENCY: CML_DATA_TYPES.DECIMAL,
